Add meta file for Galaxy compatibilty.
This change introduces a meta/main.yml file that provides metadata to Ansible Galaxy. The goal is to enable on-demand distribution of this content to anybody that's using Ansible Galaxy without affecting licensing of the content. Note that it still specifies the Apache v2 license in the metadata and will not be distributed with Ansible on installation.